Move math warping functions to utils.py
